Last version

+++ TcpServer/Command.cs
@@ -3,5 +3,5 @@ public class Command
 {
     public string ProcessName {  get; set; }
     public string CommandType { get; set; }
-    public short ProcessId { get; set; } = -1;
+    public int ProcessId { get; set; } = -1;
 }
